backend/pyproject.toml
[project]
name = "VocaStar"
version = "0.2.0"
description = "A simple Python backend implementation to provide services for VocaStar"
authors = [
    { name = "Muika(@Moemu)", email = "i@snowy.moe" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "colorlog>=6.9.0",
    "cryptography>=43.0.0",
    "fastapi>=0.116.2",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "openai>=2.2.0",
]

[project.optional-dependencies]
test = [
    "coverage>=7.10.6",
    "coverage-badge>=1.1.2",
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.2.0",
    "gevent==25.9.1",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    "ignore::DeprecationWarning:pkg_resources.*",
    "ignore:pkg_resources is deprecated as an API:UserWarning:passlib.pwd",
]

[tool.coverage.run]
concurrency = ["gevent"]
omit = [
    "app/deps/sql.py",
    "app/core/sql.py",
    "app/main.py",
    "tests/*",
]

[tool.coverage.report]
exclude_also = [
    "logger.warning",
    "logging.warning",
    "raise HTTPException",
    "raise credentials_exception",
    "except IntegrityError",
]
